Move simulation error display from Settings to Diagnostics tab
The simulation runtime error (with "Edit Item" navigation button) now
appears at the top of the Diagnostics tab instead of at the bottom of
Simulation Settings. This keeps all error information in one place.
